Classe stockage créée pour les methodes appellant des bases de données externes
