+++ app/src/main/java/com/github/wnder/MainActivity.java
@@ -30,100 +30,12 @@ import java.util.Map;
 
 public class MainActivity extends AppCompatActivity {
 
-    private FirebaseFirestore db;
-    private FirebaseStorage storage;
-    private StorageReference storageRef;
+    private Storage storage;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        this.db = FirebaseFirestore.getInstance();
-        this.storage = FirebaseStorage.getInstance();
-        this.storageRef = storage.getReference();
-
-        uploadToFirestore();
-        downloadFromFirestore();
-    }
-
-    private void uploadToCloudStorage(String filepath){
-        Uri file = Uri.fromFile(new File(filepath));
-
-        StorageMetadata metadata = new StorageMetadata.Builder()
-                .setContentType("image/jpeg")
-                .build();
-
-        UploadTask uploadTask = storageRef.child(filepath).putFile(file, metadata);
-
-        uploadTask.addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
-            @Override
-            public void onProgress(@NonNull UploadTask.TaskSnapshot snapshot) {
-                double progress = (100.0 * snapshot.getBytesTransferred()) / snapshot.getTotalByteCount();
-                System.out.println("Upload is " + progress + "% done");
-            }
-        }).addOnPausedListener(new OnPausedListener<UploadTask.TaskSnapshot>() {
-            @Override
-            public void onPaused(@NonNull UploadTask.TaskSnapshot snapshot) {
-                System.out.println("Upload is paused");
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception e) {
-                System.out.println("Upload failed");
-            }
-        }).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-            @Override
-            public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                System.out.println("Upload Succeeded");
-            }
-        });
-    }
-
-    private void downloadFromCloudStorage(String filepath) throws IOException {
-        StorageReference ref = storageRef.child(filepath);
-
-        File localFile =  File.createTempFile("images", "jpg");
-
-        ref.getFile(localFile).addOnSuccessListener(new OnSuccessListener<FileDownloadTask.TaskSnapshot>() {
-            @Override
-            public void onSuccess(FileDownloadTask.TaskSnapshot taskSnapshot) {
-                System.out.println("Download succeeded");
-            }
-        }).addOnFailureListener(new OnFailureListener() {
-            @Override
-            public void onFailure(@NonNull Exception e) {
-                System.out.println("Download Failed");
-            }
-        });
-    }
-
-    private void uploadToFirestore(){
-        Map<String, String> testMap = new HashMap<>();
-        testMap.put("Jeremy", "Jus d'orange");
-        testMap.put("Léonard", "Lasagne");
-        testMap.put("Romain", "Pizza du jeudi soir");
-        testMap.put("Nico", "Cookies");
-        testMap.put("Alois", "merci MV");
-
-
-        db.collection("test").document("doc").set(testMap);
-    }
-
-    private void downloadFromFirestore(){
-        db.collection("test")
-                .get()
-                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
-                    @Override
-                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        if (task.isSuccessful()) {
-                            for (QueryDocumentSnapshot document : task.getResult()) {
-                                System.out.println(document.getId() + " => " + document.getData());
-                            }
-                        } else {
-                            Log.w("Error", "Error getting documents.", task.getException());
-                        }
-                    }
-                });
+        storage = new Storage();
     }
 }
