+++ app/src/main/java/com/github/wnder/Storage.java
@@ -1,0 +1,121 @@
+package com.github.wnder;
+
+import android.net.Uri;
+import android.util.Log;
+
+import androidx.annotation.NonNull;
+
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.OnFailureListener;
+import com.google.android.gms.tasks.OnSuccessListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QueryDocumentSnapshot;
+import com.google.firebase.firestore.QuerySnapshot;
+import com.google.firebase.storage.FileDownloadTask;
+import com.google.firebase.storage.FirebaseStorage;
+import com.google.firebase.storage.OnPausedListener;
+import com.google.firebase.storage.OnProgressListener;
+import com.google.firebase.storage.StorageMetadata;
+import com.google.firebase.storage.StorageReference;
+import com.google.firebase.storage.UploadTask;
+
+import java.io.File;
+import java.io.IOException;
+import java.util.HashMap;
+import java.util.Map;
+
+public class Storage {
+    private FirebaseFirestore db;
+    private FirebaseStorage storage;
+    private StorageReference storageRef;
+
+    public Storage(){
+        this.db = FirebaseFirestore.getInstance();
+        this.storage = FirebaseStorage.getInstance();
+        this.storageRef = storage.getReference();
+
+        uploadToFirestore();
+        downloadFromFirestore();
+    }
+
+    private void uploadToCloudStorage(String filepath){
+        Uri file = Uri.fromFile(new File(filepath));
+
+        StorageMetadata metadata = new StorageMetadata.Builder()
+                .setContentType("image/jpeg")
+                .build();
+
+        UploadTask uploadTask = storageRef.child(filepath).putFile(file, metadata);
+
+        uploadTask.addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
+            @Override
+            public void onProgress(@NonNull UploadTask.TaskSnapshot snapshot) {
+                double progress = (100.0 * snapshot.getBytesTransferred()) / snapshot.getTotalByteCount();
+                System.out.println("Upload is " + progress + "% done");
+            }
+        }).addOnPausedListener(new OnPausedListener<UploadTask.TaskSnapshot>() {
+            @Override
+            public void onPaused(@NonNull UploadTask.TaskSnapshot snapshot) {
+                System.out.println("Upload is paused");
+            }
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception e) {
+                System.out.println("Upload failed");
+            }
+        }).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
+            @Override
+            public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
+                System.out.println("Upload Succeeded");
+            }
+        });
+    }
+
+    private void downloadFromCloudStorage(String filepath) throws IOException {
+        StorageReference ref = storageRef.child(filepath);
+
+        File localFile =  File.createTempFile("images", "jpg");
+
+        ref.getFile(localFile).addOnSuccessListener(new OnSuccessListener<FileDownloadTask.TaskSnapshot>() {
+            @Override
+            public void onSuccess(FileDownloadTask.TaskSnapshot taskSnapshot) {
+                System.out.println("Download succeeded");
+            }
+        }).addOnFailureListener(new OnFailureListener() {
+            @Override
+            public void onFailure(@NonNull Exception e) {
+                System.out.println("Download Failed");
+            }
+        });
+    }
+
+    private void uploadToFirestore(){
+        Map<String, String> testMap = new HashMap<>();
+        testMap.put("Jeremy", "Jus d'orange");
+        testMap.put("Léonard", "Lasagne");
+        testMap.put("Romain", "Pizza du jeudi soir");
+        testMap.put("Nico", "Cookies");
+        testMap.put("Alois", "merci MV");
+
+
+        db.collection("test").document("doc").set(testMap);
+    }
+
+    private void downloadFromFirestore(){
+        db.collection("test")
+                .get()
+                .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            for (QueryDocumentSnapshot document : task.getResult()) {
+                                System.out.println(document.getId() + " => " + document.getData());
+                            }
+                        } else {
+                            Log.w("Error", "Error getting documents.", task.getException());
+                        }
+                    }
+                });
+    }
+}
